establish_connection and reset blacklist
